refactor(perltidy): split native formatter contracts (#0000)
